+++ test/mocks/request_info/mocks.cc
@@ -5,9 +5,11 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
+using testing::Invoke;
 using testing::Return;
 using testing::ReturnPointee;
 using testing::ReturnRef;
+using testing::_;
 
 namespace Envoy {
 namespace RequestInfo {
@@ -33,12 +35,33 @@ MockRequestInfo::MockRequestInfo()
   ON_CALL(*this, lastDownstreamTxByteSent())
       .WillByDefault(ReturnPointee(&last_downstream_tx_byte_sent_));
   ON_CALL(*this, requestComplete()).WillByDefault(ReturnPointee(&end_time_));
+  ON_CALL(*this, setUpstreamLocalAddress(_))
+      .WillByDefault(
+          Invoke([this](const Network::Address::InstanceConstSharedPtr& upstream_local_address) {
+            this->upstream_local_address_ = upstream_local_address;
+          }));
   ON_CALL(*this, upstreamLocalAddress()).WillByDefault(ReturnRef(upstream_local_address_));
+  ON_CALL(*this, setDownstreamLocalAddress(_))
+      .WillByDefault(
+          Invoke([this](const Network::Address::InstanceConstSharedPtr& downstream_local_address) {
+            this->downstream_local_address_ = downstream_local_address;
+          }));
   ON_CALL(*this, downstreamLocalAddress()).WillByDefault(ReturnRef(downstream_local_address_));
+  ON_CALL(*this, setDownstreamRemoteAddress(_))
+      .WillByDefault(
+          Invoke([this](const Network::Address::InstanceConstSharedPtr& downstream_remote_address) {
+            this->downstream_remote_address_ = downstream_remote_address;
+          }));
   ON_CALL(*this, downstreamRemoteAddress()).WillByDefault(ReturnRef(downstream_remote_address_));
   ON_CALL(*this, protocol()).WillByDefault(ReturnPointee(&protocol_));
   ON_CALL(*this, responseCode()).WillByDefault(ReturnPointee(&response_code_));
+  ON_CALL(*this, addBytesReceived(_)).WillByDefault(Invoke([this](uint64_t bytes_received) {
+    this->bytes_received_ += bytes_received;
+  }));
   ON_CALL(*this, bytesReceived()).WillByDefault(ReturnPointee(&bytes_received_));
+  ON_CALL(*this, addBytesSent(_)).WillByDefault(Invoke([this](uint64_t bytes_sent) {
+    this->bytes_sent_ += bytes_sent;
+  }));
   ON_CALL(*this, bytesSent()).WillByDefault(ReturnPointee(&bytes_sent_));
   ON_CALL(*this, dynamicMetadata()).WillByDefault(ReturnRef(metadata_));
 }